aexpect: Use internal state to avoid bug closing Spawn objects multiple times
Whenever we .close() a Spawn() object, we're going to
clean up its base directory, removing all traces of that
object on the filesystem. Some times, though, we might
call .close() multiple times, creating a situation when
we try to recover files that were already deleted. Fix
that by managing whether objects were closed by using
a new attribute, .closed, that is set to True while it's
not closed for the first time. After the first close, all
subsequent calls to .close() will be no-ops, avoiding the
bug and keeping the temporary files created by aexpect
non predictable and secure.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>