aexpect.client: Add support to sendcontrol characters
From time to time one needs to send ctrl+... characters to the
underlying process. One can either find and use the actual value, or
they can use this function, which does the mapping itself. To use it
simply specify the key to be pressed together with ctrl and you're set.

Note: Don't mistake this method with send_ctrl, which is used to send
control messages to the aexpect process and not to your underlying
process.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>