SPEC file: rename package name to match downstream
This is one of those cases were it makes sense to track what's in
downstream.  The reason is that the SPEC file here is primarily
developed and tested under Fedora and EL.  Then, the resulting
packages are primarily run under those distros too.

IMHO it makes sense to reduce the amount of differences from
downstream to upstream, so that any changes here can be quickly picked
up by downstream and then made quickly available to end users.

Signed-off-by: Cleber Rosa <crosa@redhat.com>