scripts/aexpect-helper: read and write data in binary form
The client part of aexpect that writes to the aexpect-helper stdin is
already dealing with writes as bytes (strings which are then encoded).
This was done in commit 5aa7d475ca5f095057485ee61f98d989ee41872c.

But, aexpect-helper itself, was dealing with strings, which caused
TypeErrors exceptions and crashes such as:

   Traceback (most recent call last):
     File "/tmp/venv/bin/aexpect-helper", line 165, in <module>
       buffers[i] += data
   TypeError: must be str, not bytes

And those result in a aexpect client without a matching helper
process.

Signed-off-by: Cleber Rosa <crosa@redhat.com>