Encoding selection: use Python's locale module for choosing it
The current encoding selection makes reference to an environment
variable that has no meaning outside of aexpect itself, and it's not
documented to serve that purpose.  In fact, it was a misinterpretation
of a similarly named environment variable, PYTHONIOENCODING.

Let's use the Python's standard library `locale` for choosing the
default encoding.  Since that is a guess and may return None, let's
set "UTF-8" as a fallback.

Signed-off-by: Cleber Rosa <crosa@redhat.com>