Makefile: Enable (most) warning checks
The current defaults checks are:

    Pylint disabled: W,R,C,E1002,E1101,E1103,E1120,F0401,I0011
    Pylint enabled : W0611

But we are are ready to enable most warning checks. Let's do that to
avoid regressions, therefor the current checks use:

    Pylint disabled: W1618,R,C,E1002,E1101,E1103,E1120,F0401,I0011
    Pylint enabled : W0611

The "W1618" is kept as it's about controversial "from __future__ import
absolute_import". We use "pylint" to check our imports so let's keep it
masked.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>