Use select.poll for waiting on file descriptors
The currently used function for waiting on file descriptors for
reading is select(), which has the following downside (from man 2
select):

> select() can monitor only file descriptors numbers that are less
  than FD_SETSIZE; poll(2) does not have this limitation. See BUGS.

A process that spawns several instances of aexpect might trigger the
issue:

Traceback (most recent call last):
  File "/usr/lib64/python2.7/threading.py", line 812, in __bootstrap_inner
    self.run()
  File "/usr/lib64/python2.7/threading.py", line 765, in run
    self.__target(*self.__args, **self.__kwargs)
  File "/usr/lib/python2.7/site-packages/aexpect/client.py", line 584, in _tail
    r, w, x = select.select([fd], [], [], 0.05)
ValueError: filedescriptor out of range in select()

This patch replaces the usage of select() with poll(), from the same
module, which does not have the limitation in number of file
descriptors.

Note that poll() takes the timeout parameter in milliseconds whereas
select() used seconds.

Signed-off-by: Fabiano Rosas <farosas@linux.ibm.com>