Do not start tail thread if the child process is dead
It is possible (due to unpickling) to (re)start the tail thread after
Tail.close() has been called and Tail.reader_fds is empty.

This is currently handled by catching the TypeError exception raised
from select.select() when the file descriptor passed to it is None
(since reader_fds is empty).

The following patch in this series will replace select() with poll(),
which does not raise a TypeError. Instead of the possible solution of
testing for tail_pipe being None at the start of Tail._tail(), let's be more
explicit and avoid (re)starting the thread if Tail.close() has already
been called.

Signed-off-by: Fabiano Rosas <farosas@linux.ibm.com>