Add session ops utility with some facilitator functions
It is added as a main module since the utilities folder is reserved
only for utilities used by the core aexpect code or aexpect client.

The only detected linter problem was in an import done solely to
satisfy sphinx documentation links which can be reconsidered in
the future.

The session ops module is orgnized according to complexity and
is initially meant to be a complete selection of standard utilities
for file operations. However, there is lots of additional refactoring
and standardization to be desired and all current functions are only
contributed as-is based on the needs of their original authors.

Signed-off-by: Plamen Dimitrov <pdimitrov@pevogam.com>