Add _read_nonblocking function to report number of raw read chars
The `read_nonblocking` function ignores chars that can not be encoded,
but it should not stop processing the input when ignore chars arrived.
To address that and to allow external usage of such knowledge let's
introduce a new `_read_nonblocking` method which returns the number of
read raw characters along with the usual decoded string.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>