Make sure read_until_output_matches waits for all data
The `read_until_output_matches` function should wait for the match or
until the end of the time. Currently when `read_nonblocking` returns no
output it usually means the pipeline was closed, but it could as well
mean there were no decoded characters.

Let's use the new `_read_nonblocking` function and base our decision on
the number of raw chars read instead of the decoded values to avoid
finishing earlier than our timeout.

Signed-off-by: Haijiao Zhao <haizhao@redhat.com>
Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>