Include previous output in cmd_output_safe
The ShellSession.cmd_output_safe uses 0.5 inner timeout to get the
output which is not stored for following iterations. Similarly the
exceptions only use the current iteration output, leaving the previous
iteration(s) output discarded.

As this test is time-sensitive (or long) this commit introduces a
conditional selftest which is only executed when AEXPECT_TIME_SENSITIVE
is set. This should be performed at least before a release.

Reported-by: Yumei Huang yuhuang@redhat.com
Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>