Drop unexpected and unwanted re.sub string postprocessing
Using re.sub would result in some special characters like backslashes
being processed despite us providing repr() of the metaprogramming
embedded values. This is only a problem in some rare cases but as
we want any code to be included in a remote control file as-is in
its raw form, repr() is best utilized with a more standard string
replace() call.

To further illustrate what is meant here:

>>> print("b = " + repr(p["tmp"]))
b = 'C:\\'
>>> print(re.sub(r"a = .*", "b = " + repr(p["tmp"]), "a = abc"))
b = 'C:\'

Signed-off-by: Plamen Dimitrov <plamen.dimitrov@intra2net.com>