Avoid printing line when output_func not set
On two places we used to call _print_line even though self.output_func
was not set, which resulted in warnings. It would be nicer to put this
check directly to the _print_line function, but it is used inside
performance critical section so let's check that directly on the
corresponding places.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>