Centralize all output handling into the output function
To simplify the logging in general let's completely drop any log_file
functionality and simply require a single standard output_function
for all purposes. If users would like to stream the data or log it
into a file, they will provide a log function as an output_function
that implements this similarly to the way Avocado VT did. In this
way there also won't be confusion of whether to use a log file or
a function or both and there will only be one uniform approach to
session logging.

Signed-off-by: Plamen Dimitrov <plamen.dimitrov@intra2net.com>