Provide remote Pyro5 compatibility for multiple remote objects
The Pyro5.nameserver resp. Pyro4.naming are name servers that are
used for resolving names of remote objects (for easier access than
long hashes which are the unique default identifiers for these
objects).

With this change we will use the correct name server if we have Pyro4
installed locally and Pyro4 installed remotely and resp. Pyro5
installed locally and Pyro5 installed remotely. Before this change
we would make something like Pyro4/Pyro5 locally work with Pyro4
remotely. Note that the current usage is more standard (expecting
the same version) and is not a regression in functionality since
many of the Pyro5 versions already have incompatible protocol
with the latest (but no longer worked on) Pyro4 versions.

Sharing multiple remote objects at the same time is a more rarely
used case that we now cover with Pyro5 support as well. The method
we recommend remains in retrieving multiple remote objects one by
one using single remote object getter and this one remains more
manual and still requires some control file implementation.

The changes this builds upon are 8f4c5cc where we provide compatibility
between Pyro5 and Pyro4. The reason we went for this was to introduce
the minimal amount of changes to the various remote door functionality
until everything is better tested over the longer term. The idea was
and still is to "tread lightly".

Signed-off-by: Plamen Dimitrov <plamen.dimitrov@intra2net.com>