Removing the escape sequence from the console output!
The output coming from the console Terminal consists of Bracketed Paste escape sequence!
I have updated the code in a way that it will check the status output and see if the
output has this sequence, it will remove this from the output and then use only integer o/p of status command!

Before applying the patch status output was:
^[[?2004l0

After applying the patch status output is:
0

Signed-off-by: Anushree Mathur <anushree.mathur@linux.vnet.ibm.com>