Add rsync client for remote copy operations
This commit adds support for a new rsync client for remote copy
operations which is especially important now that the SCP protocol
is being deprecated.

Some functionality like copy between remotes is only available for
scp so we cannot add an rsync equivalent. Similarly copying from
a session information is not actively used so let's try to go without
it for the rsync backend.

Signed-off-by: Plamen Dimitrov <plamen.dimitrov@intra2net.com>