Prioritize the order of the recursive copy (directory) argument
The directory argument in copy operations is a lot more important
than most of the arguments before it and seems was historically
added last. Despite this being a breaking change it will make the
functions a lot more readable for the future so we decided to go
on with it.

Signed-off-by: Plamen Dimitrov <plamen.dimitrov@intra2net.com>